Fold Modals context into DepositWithdrawal context
