+++ src/components/Modals/TransactionWaiting/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const TransactionWaiting: React.FC = () => {
+  return <div>Waiting for confirmation</div>;
+};
+
+export default TransactionWaiting;
