+++ src/contexts/DepositWithdrawal/index.tsx
@@ -7,10 +7,17 @@ import useWallet from 'hooks/useWallet';
 
 import { parseInputValue, stake } from 'utils';
 
+type IModalType = '' | 'CONFIRM_DEPOSIT' | 'TRANSACTION_WAITING';
+
 export interface ContextValues {
   inputValue: string;
   onChangeInput: (amount: string) => string;
   onDeposit: () => Promise<boolean | ContractTransaction | undefined>;
+
+  modalOpen: boolean;
+  onCloseModal: () => void;
+  onSelectModal: (modalType: IModalType) => void;
+  selectedModal: IModalType;
 }
 
 export const DepositWithdrawalContext = React.createContext<ContextValues>({
@@ -19,11 +26,60 @@ export const DepositWithdrawalContext = React.createContext<ContextValues>({
     return amount;
   },
   onDeposit: async () => undefined,
+
+  modalOpen: false,
+  onCloseModal: () => {
+    return null;
+  },
+  onSelectModal: (modalType) => {
+    return modalType;
+  },
+  selectedModal: '',
 });
 
 export const DepositWithdrawalProvider: React.FC = ({ children }) => {
   const { signer }: { signer: Signer | null } = useWallet();
   const [inputValue, setInputValue] = React.useState<string>('');
+
+  const [modalOpen, setModalOpen] = React.useState<boolean>(false);
+  const [selectedModal, setSelectedModal] = React.useState<IModalType>('');
+
+  React.useEffect(() => {
+    document.body.addEventListener('keyup', function (e) {
+      if (e.key === 'Escape') {
+        onCloseModal();
+      }
+    });
+
+    return function cleanup() {
+      document.body.removeEventListener('keyup', function (e) {
+        if (e.key === 'Escape') {
+          onCloseModal();
+        }
+      });
+    };
+  }, []);
+
+  const onCloseModal = () => {
+    setModalOpen(false);
+  };
+
+  const onSelectModal = (modalType: IModalType): void => {
+    switch (modalType) {
+      case 'CONFIRM_DEPOSIT':
+        setModalOpen(true);
+        setSelectedModal(modalType);
+        break;
+      case 'TRANSACTION_WAITING':
+        setModalOpen(true);
+        setSelectedModal(modalType);
+        break;
+
+      default:
+        setModalOpen(false);
+        break;
+    }
+  };
 
   const onChangeInput = (amount: string) => {
     setInputValue(amount);
@@ -33,7 +89,14 @@ export const DepositWithdrawalProvider: React.FC = ({ children }) => {
   const onDeposit = async (): Promise<boolean | ContractTransaction> => {
     if (signer) {
       const response = await stake(parseInputValue(inputValue, 18).toString(), TOKEN_GEYSER_ADDRESS, signer);
-      console.log(response);
+      if (typeof response === 'boolean') {
+        return response;
+      } else {
+        onSelectModal('TRANSACTION_WAITING');
+        console.log('waiting');
+        await response.wait();
+        console.log('Complete');
+      }
       setInputValue('');
       return response;
     } else {
@@ -47,6 +110,11 @@ export const DepositWithdrawalProvider: React.FC = ({ children }) => {
         inputValue,
         onChangeInput,
         onDeposit,
+
+        modalOpen,
+        onCloseModal,
+        onSelectModal,
+        selectedModal,
       }}
     >
       {children}
